Tolerate GRIB message padding; read at float32 to restore memory profile
The first MRMS overwrite backfill with decimal quantization surfaced two
issues. Iowa Mesonet MRMS files from 2014-2015 append zero padding after
the GRIB message, which grib_decimal_scale_factors rejected (nulling
those hours as failed reads) while GDAL scans past it; locate messages
by their magic bytes instead. And reading at GDAL's native float64
tripled per-read transient memory (read_parallelism scales with node
CPUs), memory-killing backfill workers; read at float32 and round in
place — bit-identical to the float64 path and to gribberish's decode
because the float32 cast error is far below half a decimal quantum and
scaled values stay under 2^24 (verified on 2014-2020 source files).

+++ src/reformatters/noaa/mrms/conus_analysis_hourly/region_job.py
@@ -205,7 +205,7 @@ class NoaaMrmsRegionJob(
                     f"Expected exactly 1 band, found {reader.count} in {coord.downloaded_path}"
                 )
                 rasterio_band = 1
-            raw = reader.read(rasterio_band)
+            raw = reader.read(rasterio_band, out_dtype=np.float32)
 
         # GDAL unpacks GRIB values internally in float32 arithmetic (regardless of the
         # dtype read into) whose rounding error varies by architecture (amd64 vs arm64).
@@ -215,7 +215,7 @@ class NoaaMrmsRegionJob(
         decimal_scales = grib_decimal_scale_factors(coord.downloaded_path)
         decimal_scale = decimal_scales[rasterio_band - 1]
         np.round(raw, decimal_scale, out=raw)
-        result: ArrayFloat32 = raw.astype(np.float32)
+        result: ArrayFloat32 = raw
 
         nodata_sentinel = data_var.internal_attrs.nodata_sentinel
         if nodata_sentinel is not None:
